requirements.txt
# Core
python-telegram-bot==20.3
APScheduler==3.10.4
mysql-connector-python==8.2.0
sqlalchemy==2.0.41
alembic==1.16.2
cryptography==41.0.7
python-dotenv==1.0.0

# Data Export
pandas==2.2.0
matplotlib==3.10.3
lxml==5.1.0

# LLM
aiohttp==3.10.11  # For LLaMA 3.2 HTTP API

# Admin Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
starlette==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1  # Pinned for Python 3.13 compatibility
python-multipart==0.0.6
pydantic[email]==2.10.4
pydantic-settings==2.7.0
httpx>=0.24.1,<0.26.0  # Compatible with both telegram-bot and ollama
jinja2==3.1.2
openpyxl==3.1.2
reportlab==4.0.8

